requirements.txt
streamlit
Pillow
numpy
imagehash
opencv-python-headless
pandas